moved solutions in seperate modules for better seperation
