+++ src/main/main.go
@@ -10,6 +10,7 @@ import (
     log "github.com/obalunenko/logger"
     "github.com/urfave/cli/v2"
 
+    _ "solutions/day1"
 )
